requirements.txt
python-dotenv >= 0.10.0
requests >= 2.20
